backend/requirements.txt
Flask==3.0.2
tensorflow==2.16.1
Pillow==10.3.0
numpy==1.26.4
h5py==3.11.0
Werkzeug==3.0.2